+++ Arbie/Contracts/__init__.py
@@ -1,3 +1,4 @@
 """Contracts enable interacton with ETH contracts."""
+from Arbie.Contracts.balancer import PoolFactory  # noqa: F401
 from Arbie.Contracts.contract import Contract, ContractFactory, Network  # noqa: F401
 from Arbie.Contracts.uniswap import Factory, Pair  # noqa: F401